+++ ktlint-gradle-plugin/src/integrationTest/kotlin/com/rickbusarow/ktlint/SourceSetTest.kt
@@ -19,6 +19,7 @@
 package com.rickbusarow.ktlint
 
 import com.rickbusarow.ktlint.internal.Ansi.Companion.noAnsi
+import com.rickbusarow.ktlint.internal.createSafely
 import com.rickbusarow.ktlint.internal.remove
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldInclude
@@ -28,12 +29,200 @@ import org.junit.jupiter.api.Test
 internal class SourceSetTest : BaseGradleTest {
 
   @Test
+  fun `script tasks have explicit inputs for only script files`() = test {
+
+    buildFile {
+      """
+      plugins {
+        id("com.rickbusarow.ktlint")
+      }
+
+      val badTask by tasks.registering(SourceTask::class) {
+        source(buildFile)
+        outputs.file(projectDir.resolve("api/api.txt"))
+        onlyIf { true }
+      }
+      """
+    }
+
+    // creating an issue for KtLint to find
+    settingsFile { settingsFile.readText().trim() }
+
+    shouldSucceed(
+      "ktlintCheckGradleScripts",
+      "ktlintFormatGradleScripts",
+      "badTask",
+      "--rerun-tasks"
+    ) {
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :badTask UP-TO-DATE
+
+        > Task :ktlintFormatGradleScripts
+         file:///build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+         file:///settings.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+  }
+
+  @Test
+  fun `script tasks ignore script files in a build directory`() = test {
+
+    buildFile {
+      """
+      plugins {
+        id("com.rickbusarow.ktlint")
+      }
+      """
+    }
+
+    @Suppress("RemoveEmptyClassBody")
+    workingDir
+      .resolve("build/generated/my-plugin.gradle.kts")
+      .kotlin(
+        """
+        package com.test
+
+        class MyPlugin { }
+
+        """
+      )
+
+    shouldSucceed("ktlintCheckGradleScripts", "ktlintFormatGradleScripts") {
+
+      task(":ktlintCheckGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+      task(":ktlintFormatGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :ktlintFormatGradleScripts
+         file:///build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+  }
+
+  @Test
+  fun `script tasks ignore script files in an included build directory`() = test {
+
+    buildFile {
+      """
+      plugins {
+        id("com.rickbusarow.ktlint")
+      }
+      """
+    }
+
+    with(workingDir.resolve("build-logic")) {
+      resolve("build.gradle.kts")
+        .createSafely(buildFile.readText())
+      resolve("settings.gradle.kts")
+        .createSafely(settingsFile.readText().lineSequence().drop(1).joinToString("\n"))
+    }
+
+    settingsFile {
+      settingsFile.readText()
+        .replace(
+          "pluginManagement {",
+          "pluginManagement {\n  includeBuild(\"build-logic\")"
+        )
+    }
+
+    shouldSucceed("ktlintCheckGradleScripts", "ktlintFormatGradleScripts") {
+
+      task(":ktlintCheckGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+      task(":ktlintFormatGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :ktlintFormatGradleScripts
+         file:///build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+         file:///settings.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+
+    // ensure that tasks against the build-logic directory would find more stuff
+    shouldSucceed(
+      "-p",
+      "build-logic",
+      "ktlintCheckGradleScripts",
+      "ktlintFormatGradleScripts"
+    ) {
+
+      task(":ktlintCheckGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+      task(":ktlintFormatGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :ktlintFormatGradleScripts
+         file:///build-logic/build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+  }
+
+  @Test
+  fun `script tasks ignore script files in a sub project directory`() = test {
+
+    buildFile {
+      """
+      plugins {
+        id("com.rickbusarow.ktlint")
+      }
+      """
+    }
+
+    with(workingDir.resolve("lib")) {
+      resolve("build.gradle.kts")
+        .createSafely(buildFile.readText())
+    }
+
+    settingsFile {
+      settingsFile.readText()
+        .plus("include(\"lib\")")
+    }
+
+    shouldSucceed(":ktlintCheckGradleScripts", ":ktlintFormatGradleScripts") {
+
+      task(":ktlintCheckGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+      task(":ktlintFormatGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :ktlintFormatGradleScripts
+         file:///build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+         file:///settings.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+
+    // ensure that tasks against the lib would find more stuff
+    shouldSucceed(
+      ":lib:ktlintCheckGradleScripts",
+      ":lib:ktlintFormatGradleScripts"
+    ) {
+
+      task(":lib:ktlintCheckGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+      task(":lib:ktlintFormatGradleScripts")?.outcome shouldBe TaskOutcome.SUCCESS
+
+      output.remove(workingDir.path).noAnsi() shouldInclude """
+        > Task :lib:ktlintFormatGradleScripts
+         file:///lib/build.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
+
+        > Task :lib:ktlintCheckGradleScripts
+      """.trimIndent()
+    }
+  }
+
+  @Test
   fun `ktlintCheck lints Gradle script files`() = test {
 
     buildFile {
       """
       plugins {
-        kotlin("jvm")
         id("com.rickbusarow.ktlint")
 
       }
@@ -49,13 +238,8 @@ internal class SourceSetTest : BaseGradleTest {
 
       output.remove(workingDir.path).noAnsi() shouldInclude """
         > Task :ktlintCheckGradleScripts FAILED
-           file: build.gradle.kts
-                 RULE ID                               DETAIL                               FILE
-              X  standard:no-blank-line-before-rbrace  Unexpected blank line(s) before "}"  file:///build.gradle.kts:4:1:
-
-           file: settings.gradle.kts
-                 RULE ID                 DETAIL                             FILE
-              X  standard:final-newline  File must end with a newline (\n)  file:///settings.gradle.kts:1:1:
+         file:///build.gradle.kts:3:1: ❌ standard:no-blank-line-before-rbrace ╌ Unexpected blank line(s) before "}"
+         file:///settings.gradle.kts:1:1: ❌ standard:final-newline ╌ File must end with a newline (\n)
       """.trimIndent()
     }
   }
@@ -66,7 +250,6 @@ internal class SourceSetTest : BaseGradleTest {
     buildFile {
       """
       plugins {
-        kotlin("jvm")
         id("com.rickbusarow.ktlint")
 
       }
@@ -82,13 +265,8 @@ internal class SourceSetTest : BaseGradleTest {
 
       output.remove(workingDir.path).noAnsi() shouldInclude """
         > Task :ktlintFormatGradleScripts
-           file: build.gradle.kts
-                 RULE ID                               DETAIL                               FILE
-              ✔  standard:no-blank-line-before-rbrace  Unexpected blank line(s) before "}"  file:///build.gradle.kts:4:1:
-
-           file: settings.gradle.kts
-                 RULE ID                 DETAIL                             FILE
-              ✔  standard:final-newline  File must end with a newline (\n)  file:///settings.gradle.kts:1:1:
+         file:///build.gradle.kts:3:1: ✅ standard:no-blank-line-before-rbrace ╌ Unexpected blank line(s) before "}"
+         file:///settings.gradle.kts:1:1: ✅ standard:final-newline ╌ File must end with a newline (\n)
       """.trimIndent()
     }
   }